refactor: component design
