+++ __tests__/cookie-footer.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { CookieFooterWrapper } from '../src/components/CookieFooterWrapper'
+import { CookieFooterWrapper } from '../src/components/Footer/cookie'
 
 describe('Cookie Footer Wrapper', () => {
     it('render correctly', () => {
