+++ src/components/Card/index.tsx
@@ -2,7 +2,7 @@ import React, { FC, Fragment } from 'react'
 import { useIntl } from 'gatsby-plugin-intl'
 import * as SC from './index.module.scss'
 import LazyImage from '../LazyImage'
-import Skills from '../Skills'
+import Skills from './skills'
 import { SiteMetadataTypes } from '../../types'
 import { NekoButton } from 'nekohack-ui'
 import PaperPlaneSvg from '../../static/icons/paper-plane.svg'
