+++ src/components/Card/skills.module.scss
