+++ src/components/Footer/cookie.tsx
